+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Security: Prevent package push from publishing files outside the project root (GHSA-85m4-5f4j-mrr5)
 
 ## [1.4.2] - 2026-04-13